move series lib to shared code
