+++ src/features/currency/components/DataChart/DataChart.tsx
@@ -10,8 +10,8 @@ import {
   Tooltip,
 } from '@highcharts/react';
 import { StockChart, StockSeries } from '@highcharts/react/Stock';
+import { getSeriesColor } from '@/lib/series';
 import { EXPLORE_FRACTION_DIGITS, MS_3_MONTHS } from '../../constants';
-import { getSeriesColor } from '../../utils';
 import { type DataChartProps } from './DataChart.types';
 
 export const DataChart: FunctionComponent<DataChartProps> = ({
