--- src/features/currency/utils.ts
@@ -1,5 +1,0 @@
-import { SERIES_COLORS } from './constants';
-
-export const getSeriesColor = (seriesIndex: number): string => {
-  return SERIES_COLORS[seriesIndex % SERIES_COLORS.length];
-};
